update to the WASI 0.2.0-rc-2023-10-18 snapshot
This requires updating `wasmtime`, `wasmtime-wasi`, `wasm-compose`, the WIT
files and preview1 adapter (built from the `wasmtime` `v14.0.4` tag).

I had to rewrite a large portion of the `virtual-adapter` subcrate given how
much code the "resourcification" of WASI affected.  I did this by removing all
the code and adding modified chunks of it back in incrementally based on
compiler feedback.  This means the code is reordered somewhat -- sorry about
that.  It was bound to be a messy diff no matter what.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>