fix: truly respect `--allow-all` to allow all components (#61)
`ArgAction::SetTrue` is not the right action for `allow_xxx` options, as
they should default to `None` instead of `Some(false)` when the option
itself is not specified. The wrong default value led to the `--allow-all`
option not being respected, as the `allow_xxx.unwrap_or(allow_all)`
always ignored the `allow_all` value.