chore: remove use of nightly features in Cargo.toml
The use of `profile-rustflags` is mainly to minimize the size of
virtual-adapter, but it's declared in the root Cargo.toml since it's
only allowed in the workspace root even though it's not used when
building wasi-virt crate.

The use of the nightly feature requires projects depending on wasi-virt
crate to also use nightly, which complicates the build process
(especially building native extensions of Ruby..)

This commit moves the `rustflags` setting from the Cargo.toml to the
RUSTFLAGS env var in build-adapter.sh.