dep: Optionalize `wasm-opt` crate dependency
The `wasm-opt` crate brings in C++ dependencies (binaryen) and it makes
the build process more complex and slower (also setting up
cross-compilation for C++ is not trivial unlike Rust).
Also the DCE process doesn't make much difference if we allow all WASI
subsystems.
This commit makes the `wasm-opt` crate dependency optional so that crate
users can choose whether to use it or not.

Also strip the "name" section at walrus module generation time if
`debug=false` because it was stripped by `wasm-opt` before.