Add support for `wasi:config/runtime` virtualization
The `wasi:config/runtime` interface is a way to pass configuration to
other components with a standard interface. Similar to
`wasi:cli/environment` the core data model is a set of key-value pairs.

The proposed API and CLI support in WASI-Virt mirrors the support for
environment variables that currently exists.

The config runtime proposal is currently stage 2 and does not a have a
release.

Signed-off-by: Scott Andrews <scott@andrews.me>