Rename `wasi:config/runtime` to `wasi:config/store` (#100)
The `wasi:config/runtime` interface was renamed upstream to
`wasi:config/store`. The repository was also renamed from
github.com/WebAssembly/wasi-runtime-config to
github.com/WebAssembly/wasi-config, cementing the shift away from
"runtime".

This commit replaces the support for the runtime interface and replaces
it with the functionally equivalent store interface. It also removes
references to "runtime" config in the readme and CLI help.

Since the interface is still draft and support labeled as experimental,
backwards compatibility is not preserved.

Signed-off-by: Scott Andrews <scott@andrews.me>